Create directory to hold tab info if provided
There is now an optional positional argument that may be passed to the
program called "tab_directory", which, if provided, will create a new
directory at the given path if it doesn't already exist.

If the argument is NOT passed, no directory is created and the browser
essentially operates in "Private Browsing" mode.