Save pages to tabs
Any page the user requests now ends up in a .browsertab file if a tab
directory is passed as a commandline argument.

Right now, .browsertab files are just regular text files, but I've given
them this extension so that we KNOW for sure that they're tabs for this
browser and not just regular plaintext placed in there by other users or
developers.