Ensure resolver converts offline addresses
Since right now, this is in the early stages of the project, there's no
support for rendering web pages (I know, it's weird).

Instead, we want to ensure that URLs can be loaded from
internally-specified variables (also weird, but that's what the project
currently calls for).  With this commit, the URLs can be converted to
those variable names, but it's ONLY acting on domain names at the
moment.  This may hopefully change soon.