Ensure browser loads offline pages
This should complete the page loading feature as it stands for now.  At
the moment, it's probably been mentioned in a previous commit, the
browser will ONLY render pages whose URLs can be translated into
variables in the global namespace.  This is probably REALLY dangerous,
but when URLs start being loaded online like the would be in a proper
browser, this code can be removed.